+++ dbt/include/databricks/macros/materializations/table.sql
@@ -32,6 +32,8 @@
       {% endif %}
     {% endif %}
 
+    {% do validate_persist_doc_columns(target_relation, model) %}
+
     {% set should_revoke = should_revoke(existing_relation, full_refresh_mode=True) %}
     {{ apply_grants(target_relation, grant_config, should_revoke) }}
     {% do optimize(target_relation) %}